Fix everything needed to basic dyn probe instrum.
Fixed a few problems from the latest version: memory allocated with
mmap should be PRIVATE, marked as READ and EXECUTE only, and the file
used during allocation shouldn't be deleted. Probe firing function
was implemented in assembly (x86_64) to allow more control on what's
happening, but maybe it won't be necessary on later versions.

README updated with one-liner to instrument mainer:lorem